Add types package which defines basic VM types
This adds the `types` and `types/typeid` packages which define the
generic type IDs, interfaces for values in the VM (both plain values and
stack values), as well as implementations of those interfaces.

The interfaces provided are Value and StackValue. Value is the interface
all types must implement and only provides basic informational
functions, along with an `Upcast` function to convert the Value into the
associated StackValue. StackValue is a Value which provides a `Downcast`
function for converting to a requested `Value`. In general, Values are
stored on structs, while StackValues are the only types allowed within
the stack.

Currently only the numeric types are implemented: Uint8, Uint16, Uint32,
Uint64, Int8, Int16, Int32, Int64, Float32, and Float64. Of these types,
Float64, Int64, and Uint64 are stack values. All other types are just
Values.