Add golangci-lint configuration and fix reported errors
This adds a configuration file for the golangci-lint linter suite and
fixes any errors which the linters flagged. Not all linters were
enabled; the criteria for enabling a linter was that the linter would
check for something that the code here could trip (so no sql linters,
for example) and that no other duplicate linter was also enabled (e.g.
gofumpt vs revive).