Define opcodes, vm, and move immediates operations to vm
This defines two new packages: vm and opcodes. The opcodes package
contains the enumeration of the opcodes for the VM, while the vm package
defines the virtual machine structures and code.

Additionally, this moves the immediates fetching code from the bytecode
package to the vm package as functions on the Thread type.